+++ tests/testthat.R
@@ -7,6 +7,6 @@
 # * https://testthat.r-lib.org/reference/test_package.html#special-files
 
 library(testthat)
-library(viticolaR)
+library(viticolR)
 
-test_check("viticolaR")
+test_check("viticolR")